test(contact): Add data for pending Test Case 2
'contact-form.json' constitutes plausible data for entry into the form
on the Contact page which will be used in Test Case 2.

Test Case 2 will verify that the form on the Contact page successfully
submits when its mandatory fields are populated with valid data.

This commit also adds the 'data' directory.