test(contact): Add Test Case 2
Test Case 2 verifies that the form on the Contact page successfully
submits when its mandatory fields are populated with valid data.

The test is executed 5 times per 'Contact page: Form behaviour' test run
to ensure consistent success, fulfilling the required condition
stipulated in the brief: "Note: Run this test 5 times to ensure 100%
pass rate"

Each test execution uses a different dataset defined in the
'contact-form.json' file added in 48475c3. For example, "Abigail
Johnson" is used as the person for the first execution, and "Zachary
O'Keeffe" is used as the person for the second execution.