ci(lint): Add lint job
Utilise GitHub's SuperLinter to lint the codebase prior to executing the
Playwright tests.

A separate workflow should not be necessary for this job, so it has been
prepended to the jobs list within the Playwright workflow.

Accessing SuperLinter via Docker Hub is a simple way of always accessing
the latest version.