ci(lint): Address lint job issues
Attempt to fix error with the 'TYPESCRIPT_STANDARD' linter running
within the Super-Linter job.

The 'TYPESCRIPT_ES' linter also does not appear to be functioning as
expected, as the rules specified within the ESLint config,
eslint.config.mjs, currently result in error detection locally, but not
remotely.

Working theory is that the file is not being recognised, so the
supported Super-Linter env 'TYPESCRIPT_ES_CONFIG_FILE' has been set.